src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==1.0.3
httpx>=0.23.0,<0.24.0
sentence-transformers==2.2.2
faiss-cpu>=1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
numpy>=1.26.0
python-multipart==0.0.6
requests==2.31.0 